Migrate one-one-one test to new framework
No changes to the test content
