+++ tests/integration/query/one_to_one_to_one/simple_test.go
@@ -17,77 +17,112 @@ import (
 )
 
 func TestQueryOneToOneToOne(t *testing.T) {
-	test := testUtils.RequestTestCase{
+	test := testUtils.TestCase{
 		Description: "One-to-one-to-one relation primary direction",
-		Request: `query {
-			Author {
-				name
-				published {
-					name
-					publisher {
-						name
+		Actions: []any{
+			testUtils.SchemaUpdate{
+				Schema: `
+					type Book {
+						name: String
+						rating: Float
+						author: Author
+						publisher: Publisher @primary
 					}
-				}
-			}
-		}`,
-		Docs: map[int][]string{
-			//books
-			0: {
+
+					type Author {
+						name: String
+						age: Int
+						verified: Boolean
+						published: Book @primary
+					}
+
+					type Publisher {
+						name: String
+						address: String
+						yearOpened: Int
+						printed: Book
+					}
+				`,
+			},
+			testUtils.CreateDoc{
+				CollectionID: 0,
 				// "bae-a6cdabfc-17dd-5662-b213-c596ee4c3292"
-				`{
+				Doc: `{
 					"name": "Painted House",
 					"publisher_id": "bae-1f4cc394-08a8-5825-87b9-b02de2f25f7d"
 				}`,
+			},
+			testUtils.CreateDoc{
+				CollectionID: 0,
 				// "bae-bc198c5f-6238-5b50-8072-68dec9c7a16b"
-				`{
+				Doc: `{
 					"name": "Theif Lord",
 					"publisher_id": "bae-a3cd6fac-13c0-5c8f-970b-0ce7abbb49a5"
 				}`,
 			},
-			//authors
-			1: {
-				`{
+			testUtils.CreateDoc{
+				CollectionID: 1,
+				Doc: `{
 					"name": "John Grisham",
 					"published_id": "bae-a6cdabfc-17dd-5662-b213-c596ee4c3292"
 				}`,
-				`{
+			},
+			testUtils.CreateDoc{
+				CollectionID: 1,
+				Doc: `{
 					"name": "Cornelia Funke",
 					"published_id": "bae-bc198c5f-6238-5b50-8072-68dec9c7a16b"
 				}`,
 			},
-			// publishers
-			2: {
+			testUtils.CreateDoc{
+				CollectionID: 2,
 				// "bae-1f4cc394-08a8-5825-87b9-b02de2f25f7d"
-				`{
+				Doc: `{
 					"name": "Old Publisher"
 				}`,
+			},
+			testUtils.CreateDoc{
+				CollectionID: 2,
 				// "bae-a3cd6fac-13c0-5c8f-970b-0ce7abbb49a5"
-				`{
+				Doc: `{
 					"name": "New Publisher"
 				}`,
 			},
-		},
-		Results: []map[string]any{
-			{
-				"name": "John Grisham",
-				"published": map[string]any{
-					"name": "Painted House",
-					"publisher": map[string]any{
-						"name": "Old Publisher",
+			testUtils.Request{
+				Request: `query {
+					Author {
+						name
+						published {
+							name
+							publisher {
+								name
+							}
+						}
+					}
+				}`,
+				Results: []map[string]any{
+					{
+						"name": "John Grisham",
+						"published": map[string]any{
+							"name": "Painted House",
+							"publisher": map[string]any{
+								"name": "Old Publisher",
+							},
+						},
 					},
-				},
-			},
-			{
-				"name": "Cornelia Funke",
-				"published": map[string]any{
-					"name": "Theif Lord",
-					"publisher": map[string]any{
-						"name": "New Publisher",
+					{
+						"name": "Cornelia Funke",
+						"published": map[string]any{
+							"name": "Theif Lord",
+							"publisher": map[string]any{
+								"name": "New Publisher",
+							},
+						},
 					},
 				},
 			},
 		},
 	}
 
-	executeTestCase(t, test)
+	testUtils.ExecuteTestCase(t, []string{"Book", "Author", "Publisher"}, test)
 }
