--- tests/integration/query/one_to_one_to_one/utils.go
@@ -1,44 +1,0 @@
-// Copyright 2022 Democratized Data Foundation
-//
-// Use of this software is governed by the Business Source License
-// included in the file licenses/BSL.txt.
-//
-// As of the Change Date specified in that file, in accordance with
-// the Business Source License, use of this software will be governed
-// by the Apache License, Version 2.0, included in the file
-// licenses/APL.txt.
-
-package one_to_one_to_one
-
-import (
-	"testing"
-
-	testUtils "github.com/sourcenetwork/defradb/tests/integration"
-)
-
-var bookAuthorGQLSchema = `
-	type Book {
-		name: String
-		rating: Float
-		author: Author
-		publisher: Publisher @primary
-	}
-
-	type Author {
-		name: String
-		age: Int
-		verified: Boolean
-		published: Book @primary
-	}
-
-	type Publisher {
-		name: String
-		address: String
-		yearOpened: Int
-		printed: Book
-	}
-`
-
-func executeTestCase(t *testing.T, test testUtils.RequestTestCase) {
-	testUtils.ExecuteRequestTestCase(t, bookAuthorGQLSchema, []string{"Book", "Author", "Publisher"}, test)
-}
